requirements.txt
# Core Flask
Flask==2.3.2
Werkzeug==2.3.6
Jinja2==3.1.2

# Email & Forms
Flask-Mail==0.9.1
Flask-WTF==1.1.1
WTForms==3.0.1
email-validator==2.0.0

# Google Sheets API
gspread==5.7.2
oauth2client==4.1.3

# Environment & Deployment
python-dotenv==1.0.0
gunicorn==20.1.0

# Database (if needed later)
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.4
psycopg2-binary==2.9.6

# Utilities
Pillow==10.0.0       # For image processing
requests==2.31.0      # For HTTP requests